app/src/requirements.txt
django
gunicorn
djangorestframework
python-decouple
psycopg2-binary==2.9.1
praw
prawcore
